% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioDat.R
\docType{data}
\name{bioDat}
\alias{bioDat}
\title{plot-level data: above and belowground biomass}
\format{A dataframe with the following columns:
\describe{
\item{yr}{year of sample collection}
\item{mo}{month of sample collection (3-character form)}
\item{mo2}{month of sample collection (numeric)}
\item{region}{region}
\item{site}{name of sampling marsh}
\item{lat}{latitude, decimal degrees}
\item{long}{longitude, decimal degrees}
\item{plot}{plot name (A, B, C)}
\item{stems}{live stem density. Units: stems per square meter}
\item{lngth.median}{median live stem length. Units = centimenters}
\item{mass}{total live biomass in plot. Units = grams per square meter}
\item{stems.dead}{dead stem density. Units: stems per square meter}
\item{lngth.median.dead}{median dead stem length. Units = centimenters}
\item{mass.dead}{total dead standing biomass in plot. Units = grams per square meter}
\item{litter}{total litter biomass in plot. Units = grams per square meter}
\item{live.bg.gm2}{live belowground biomass to a depth of 30 cm. Units = grams per square meter}
\item{dead.bg.gm2}{dead belowground biomass to a depth of 30 cm. Units = grams per square meter}
\item{tot.bg.gm2}{total belowground biomass to a depth of 30 cm. Units = grams per square meter}
}}
\description{
plot-level data: above and belowground biomass
}
\keyword{datasets}
