% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getParams}
\alias{getParams}
\title{Generates allometry models for input data}
\usage{
getParams(dataset, massCol = "mass_tot", heightCol = "hgt",
  typeCol = "status", cutOff = 5, startVals = 0.2, timeName = NA,
  dataName = NA, returnPlot = FALSE, savePlot = FALSE,
  plotTitle = NA)
}
\arguments{
\item{dataset}{input dataframe}

\item{massCol}{name of column with mass data}

\item{heightCol}{name of column with height data}

\item{typeCol}{name of column with live/dead status}

\item{cutOff}{minimum number of stems to be used for allometry model. Smaller numbers increase odds of convergence problems.}

\item{startVals}{starting value used for allometry model coefficients and exponents}

\item{timeName}{optional label for time period}

\item{dataName}{optional label for data/site}

\item{returnPlot}{if TRUE, a plot of allometry data is returned. Two panels, with live stems are on the left, dead are on right}

\item{savePlot}{if TRUE, plot is saved to working directory}

\item{plotTitle}{optional title for plot}
}
\value{
value
}
\description{
Generates allometry models for input data
}
\examples{
getParams
}
