% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\docType{data}
\name{ndvi}
\alias{ndvi}
\title{plot-level data: normalized difference vegetation index}
\format{A dataframe the following variables:
\describe{
\item{yr}{}
\item{site}{}
\item{mo}{}
\item{avg.ndvi}{}
\item{sd.ndvi}{}
}}
\description{
plot-level data: normalized difference vegetation index
}
\keyword{datasets}
