% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutDat.R
\docType{data}
\name{nutDat}
\alias{nutDat}
\title{plot-level data: soil and water nutrients and physical parameters}
\format{A dataframe}
\description{
plot-level data: soil and water nutrients and physical parameters
}
\keyword{datasets}
