% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDat.R
\docType{data}
\name{plotDat}
\alias{plotDat}
\title{plot-level data: biomass, soil, and water nutrient and physical parameters}
\format{A dataframe the following variables:
\describe{
\item{year}{year of sample collection}
\item{month.name}{month of sample collection (3-character form)}
\item{month.number}{month of sample collection (numeric)}
\item{region}{region}
\item{site}{name of sampling marsh}
\item{lat}{latitude, decimal degrees}
\item{long}{longitude, decimal degrees}
\item{dist}{distance from marsh edge, meters}
\item{plot}{plot name (A, B, C)}
\item{stems}{live stem density. Units: stems per square meter}
\item{lngth.median}{median live stem length. Units = centimenters}
\item{mass}{total live biomass in plot. Units = grams per square meter}
\item{stems.dead}{dead stem density. Units: stems per square meter}
\item{lngth.median.dead}{median dead stem length. Units = centimenters}
\item{mass.dead}{total dead standing biomass in plot. Units = grams per square meter}
\item{litter}{total litter biomass in plot. Units = grams per square meter}
\item{live.bg.gm2}{live belowground biomass to a depth of 30 cm. Units = grams per square meter}
\item{dead.bg.gm2}{dead belowground biomass to a depth of 30 cm. Units = grams per square meter}
\item{tot.bg.gm2}{total belowground biomass to a depth of 30 cm. Units = grams per square meter}
\item{temp.soil}{soil temperature in degrees C}
\item{ph.soil}{soil pH in pH units}
\item{bulk.density.g.cm3.soil}{soil bulk density in grams per cubic centimeter}
\item{grav.water.content.soil}{soil water content. Units = grams/gram dry mass}
\item{om.soil}{organic matter content of soil collected in top 5 cm, measured by loss on ignition. Units = grams/gram}
\item{c.soil}{organic carbon content of soil collected in top 5 cm and fumigated with HCl. Units = grams/gram}
\item{tn.soil}{total nitrogen content of soil collected in top 5 cm and fumigated with HCl. Units = grams/gram}
\item{tp.ppm.soil}{total phosphorous content of soil collected in top 5 cm and fumigated with HCl. Units = milligrams/gram}
\item{extractable.no3.mmol.gdw.soil}{extractable NO3+NO2 in soil. Units = millimol N per gram dry mass}
\item{extractable.nh4.mmol.gdw.soil}{extractable NH4 in soil. Units = millimol N per gram dry mass}
\item{extractable.po4.mmol.gdw.soil}{extractable PO4 in soil. Units = millimol P per gram dry mass}
\item{porewater.salinity.soil}{porewater salinity. Units = psu}
\item{redox.soil}{soil redox. Units = milliVolts}
\item{water.temp.bay}{temperature of bay water adjacent to site. Units = degrees C}
\item{salinity.psu.bay}{salinity of bay water adjacent to site. Units = psu}
\item{sp.cond.bay}{specific conductivity of  bay water adjacent to site. Units = mS/cm}
\item{dissolved.no3.no2.bay}{dissolved NO3+NO2 in bay water adjacent to site. Units = millimoles N/L}
\item{dissolved.nh4.bay}{dissolved NH4 in bay water adjacent to site. Units = millimoles N/L}
\item{dissolved.po4.bay}{dissolved PO4 in bay water adjacent to site. Units = millimoles P/L}
\item{dissolved.sio2.bay}{dissolved SiO2 in bay water adjacent to site. Units = millimoles Si/L}
\item{water.depth.cm.}{depth of overlying water at marsh sampling location. Units = centimeters}
\item{water.temp}{temperature of overlying water at marsh sampling location. Units = degrees Celsius}
\item{salinity.psu.}{salinity of overlying water at marsh sampling location. Units = psu}
\item{sp.cond}{specific conductivity of overlying water at marsh sampling location. Units = mS/cm}
\item{dissolved.no3.no2}{dissolved NO3+NO2 in overlying water at marsh sampling location. Units = millimoles N/L}
\item{dissolved.nh4}{dissolved NH4 in overlying water at marsh sampling location. Units = millimoles N/L}
\item{dissolved.po4}{dissolved PO4 in overlying water at marsh sampling location. Units = millimoles P/L}
\item{dissolved.sio2}{dissolved SiO2 in overlying water at marsh sampling location. Units = millimoles Si/L}
}}
\description{
plot-level data: biomass, soil, and water nutrient and physical parameters
}
\keyword{datasets}
