% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonLabel.R
\name{seasonLabel}
\alias{seasonLabel}
\title{Labels seasons}
\usage{
seasonLabel(data, monthYearColumn = "moYr", siteColumn = "site",
  seasons = NA, year = c(13:18))
}
\arguments{
\item{data}{input dataframe}

\item{monthYearColumn}{character element identifying month/year column. Column data should be in the format mmm-yy}

\item{siteColumn}{character element identifying site identifier column}

\item{seasons}{a list associating each month with a season}

\item{year}{vector of two-digit years}
}
\value{
adds columns to input dataframe
}
\description{
Labels seasons
}
\examples{
seasonLabel

}
